\name{cudia-internal}
\title{Internal cudia functions}
\alias{cudia.default}
\alias{cudiaDet}
\description{
Estimate the CUDIA model parameters, then output cross-level imputed values.
A wrapper of cudiaDet.
}
\details{These are not intended for use by users. \code{cudia.default} and \code{cudiaDet} implement the Bregman deterministic clustering algorithm.}
\usage{
cudiaDet(formula, data, K)
}
\author{Yubin Park}
\keyword{ internal }
