\name{cudia_medic}
\alias{cudia_medic}
\title{Artificial medicare state-level spendings with individual demographic features}
\description{
 The \code{cudia_medic} data frame as 5 columns: 4 individual-level and 1 aggregate-level medicare spendings.}
\usage{
cudia_medic
}
\format{
This data frame contains the following columns:
\describe{
\item{income}{individual-level income data}
\item{house_value}{value of each individual's house}
\item{age}{age of each individual}
\item{se_index}{socio-economic index of each individual}
\item{medicare_d}{state-level medicare summary (durable goods)}
}
}
\source{
J. Ghosh and Y. Park, \emph{Integrated Predictive Modeling of Variable-Resolution Healthcare Data}, Meaningful Use of Complex Medical Data 2012.
}
\keyword{ datasets }


